OpenSea/requirements.txt
git+https://github.com/Rapptz/discord.py
requests
discord-ext-bot
pytz
datetime
simplejson
python-dotenv